mp4muxer: Convert duration to timescale based value.
The Mp4MuxerDelegate provides TimeTicks for the 'trun' box's
sample duration and input TimeTicks based duration for the
'tfra' entry time.

The change is to convert these Time's millisecond into the MP4
box ('mdhd') timescale based on the spec.

It also updated relevant tests.

Bug: 1471314
Bug: 1465031
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4906154
Commit-Queue: Sunggook Chue <sunggch@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1205613}

+++ media/muxers/mp4_muxer_delegate.h
@@ -99,8 +99,8 @@ class MEDIA_EXPORT Mp4MuxerDelegate {
 
   // video and audio index is a 0 based index that is an item of the container.
   // The track id would be plus one on this index value.
-  absl::optional<int> video_track_index_;
-  absl::optional<int> audio_track_index_;
+  absl::optional<size_t> video_track_index_;
+  absl::optional<size_t> audio_track_index_;
   int next_track_index_ = 0;
 
   // Duration time delta for the video track.
